sealguard-api/pyproject.toml
[project]
name = "sealguard-api"
version = "0.1.0"
description = "SealGuard backend API"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "pydantic==2.9.2",
  "sqlalchemy==2.0.35",
  "python-multipart==0.0.9",
  "python-dotenv==1.0.1",
  "psycopg2-binary==2.9.9",
  "ultralytics==8.3.26",
  "pyyaml==6.0.2",
  "opencv-python-headless==4.10.0.84",
]

[tool.uv]
package = false
